+++ src/Exception/Finvoice/OperatorCodeHasSpacesException.php
@@ -1,0 +1,4 @@
+<?php
+namespace Laskuhari\Exception\Finvoice;
+
+class OperatorCodeHasSpacesException extends FinvoiceException {}