+++ src/Exception/Finvoice/OperatorCodeMismatchException.php
@@ -1,0 +1,4 @@
+<?php
+namespace Laskuhari\Exception\Finvoice;
+
+class OperatorCodeMismatchException extends FinvoiceException {}